Refactored test descriptor execution paradigm
