+++ tests/src/test/java/org/antublue/verifyica/test/scenario/ScenarioTest3.java
@@ -75,7 +75,7 @@ public class ScenarioTest3 {
                 argumentContext.getTestArgument(), argumentContext.getTestArgument().getPayload());
     }
 
-    @Verifyica.Test
+    @Verifyica.Test(order = 7)
     public void random(ArgumentContext argumentContext) throws Throwable {
         System.out.printf(
                 "random(name[%s], payload[%s])%n",
